Refactor: Change Evaluation id type to Long
- Changed the Evaluation entity's id type from Int to Long.
- Updated the EvaluationDao to return Long from the insert method and to accept Long for the delete method.
- Updated the EvaluationRepository to return Long from the insert method and to accept Long for the delete method.
- Updated the EvaluationViewModel to use liveData for the insert method and to accept Long for the delete method.
- Updated the UI to use Long for the id of evaluations.
- Updated the navigation arguments to use Long for the id of evaluations.
- Updated the UI to reflect the changes in id type.

+++ app/src/main/java/com/firmino/hinedigital/model/entity/Evaluation.kt
@@ -10,7 +10,7 @@ import com.firmino.hinedigital.model.converters.ConverterStringArray
 @Entity(tableName = "evaluation_table")
 @TypeConverters(ConverterIntArray::class, ConverterStringArray::class)
 data class Evaluation(
-    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "id") val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) @ColumnInfo(name = "id") val id: Long = 0L,
     @ColumnInfo(name = "name") val name: String,
     @ColumnInfo(name = "gender") val gender: String,
     @ColumnInfo(name = "birthday") val birthday: String,
